refatoraçao do codigo
